(BWDIC!) Obsolete *sendmail*++ in favour of *mta*++; obsolete *smtp*..
It is highly disturbing to go *sendmail* but talk about MTA.
Let's change that.  Introduce *mta*, *mta-argv0*, etc., obsolete
*sendmail* stuff.

And then it must be said that i dislike the method.
Make it possible to assign an URL to *mta*, and if it is a SMTP
one, well, then let's talk SMTP instead.
Therefore obsolete *smtp*, too.  Provide some small hacks to
support (read: _ignore_) file:// URLs where necessary.
All that for real, hopefully, in v15.  Hopefully.

This is BackWardInCompatible because of the changed configuration
settings (VAL_SENDMAIL -> VAL_MTA, VAL_SENDMAIL_PROGNAME ->
VAL_MTA_ARGV0).

Yes i know it is odd to have *mta-argv0* which only applies to
file-based *mta*, but *smtp-hostname*, which applies to SMTP based
*mta*.  In TODO we know have an idea to say
  mta=smtp://a:b@xy.z?no-starttls;hostname=bla
But let's see what the future brings