extract_date_from_from_(): dig more invalid MBOXes (Antonio Radici)..
Antonio Radici (antonio at debian dot org) reopened a Debian bug
and forwarded it to the mutt bug list.

  https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=624111

So let's just support more "invalid" From_ lines with pseudo-
obfuscated addresses.
TODO We need to generalize this somewhat; do not simply skipword()
TODO but add a set of supported things, just like for cmatch() a.k.a.
TODO the date; and do be very laxe and try very hard if !*mbox-rfc4155*