sendout.c: fix HAVE_DEBUG From_ quoting..
Rename __savemail() to a_sendout__savemail() and rewrite it a bit.

If we resend a message then we (at least: yet) don't reevaluate
it's contents for real, so if a user resends a message that was
produced by a MUA which doesn't force MIME encoding when there is
"From " in a message body, then (with *record-resent*) we need to
MBOXO quote the message if we save it in *record*.
Sofar i had a HAVE_DEBUG assertion which did _not_ do it, but the
regular user shipout code always did.

Also there was a resource leak in the error path.  The file would
have been cleaned up when the command loop ticks, though.