popen.c:_file_save(): FIX false cherry-pick [41250b1] for IMAP..
This is false cherry-pick back to [master].
Of course the changed fseek(3) behaviour should have been adjusted
to also apply to the IMAP code that is still included in the v14.8
branch (and maybe will be again in v14.9, but not so at the time
of this writing).

The result of this messed cherry-pick is that things like
"copy * IMAP-URL" don't work in v14.8.10, because the file is at
a false offset!  Fetching still works...