is_head(): change compat argument to check_rfc4155..
Whereas *mbox-rfc4155* was a no-brainer and we should really
simply dig any MBOX as long as it uses "From " From_ lines as of
POSIX, at least *debug*/*verbose* warn users if a MBOX uses From_
lines which don't comply to RFC 4155.  Do so only once per
mailbox, to not flood users which open freedesktop.org or so
archives (they obfuscate addresses even in From_ lines).