Change n_iconv*() protos, add enum n_iconv_flags (Ralph Corderoy)..
On FreeBSD and Solaris the tests currently fail because their
iconv(3) happily opens a UTF8->ASCII conversion and then we we're
not prepared to handle non-reversible replacements as
non-ignorable errors, as Ralph has pointed out correctly.

Add a flag enum so that we now can treat any replacement as an
error, too.

While doing this is again saw quite a lot of '?' replacement
characters we put, makeprint() uses Unicode if applicable.  We
can't control the non-reversible of iconv(3).  Hoping for S-CText.