FIX [b9e4e61d] (Drop features', change `version'..)..
The initial idea was different to avoid the parsing of the feature
string, but the final version did so.  However, we still copied
over another strlen() full of data after the first NUL, which
obviously is a bad idea.  Interesting that this wasn't yet
catched by any SEGFAULT or whatever!