Add *typescript-mode*, set it for -# (Ralph Corderoy, Martin Neitzel)..
Ralph ran us via script(1), and for us this looks like a truly
interactive session, so we do use colours, enable the line editor,
etc., which garbles the generated typescript somewhat.
Martin jumped in regarding -#.

So instead of requiring the user to know so much and having a need
to adjust the code that deals with -# due to this, simply invent
a new VIP multiplexer variable *typescript-mode* which adjusts all
knobs which somehow could interfere with running us in things like
script(1), set it for -# by default.  It doesn't restore any
former setting when unset.