Add *inbox* (Stephen Isard, and Jürgen Daubert)..
stable/v14.8:
  After a long thread, mostly in private, it seems to be the time
  NOW to introduce a bypass to avoid that setting *folder* to an
  IMAP mailbox causes automatic connection to apply, with an
  implicit "INBOX" mailbox.

  So temporarily keep that for compatibility, but prefer *inbox*
  when looking for %.

For v14.9, drop that idea of pimping $MAIL that has been
introduced in [e337e00] (Support expansions in $MAIL!..) again:
$MAIL is standardized by POSIX as an absolute pathname, so
allowing expansions in there is a messy thing to do.  This is even
more true given that $MAIL is automatically exported to child
processes, and passing MAIL=$'%:+${this}' to a child is hardly
ever something useful to do.