folder_query(): err: real FIX for problem "fixed" with [6a09b2c]..
No, the resulting strlen() cannot be 0 there, that was logically
correct.  The real problem was that in the #ifdef
HAVE_REALPATH_NULL branch we used n_string_assign_cp(), but which
-- different to n_str_assign_cp()! -- does not NUL terminate the
buffer, so that the following strlen() went to knowhere.

The problem why [6a09b2c] fixed the things for me is that tcc(1)
doesn't execute this branch since it will not detect a realpath(3)
that can take a NULL buffer during configuration time, thanks to
Michael Matz of tcc(1) who described this in detail in
alpine.LSU.2.20.1606060151180.9036@hell6.fritz.box, it is about
symbol versioning which tcc(1) doesn't support.