`bind': new "default" context and changed "base" semantics..
We really need to differentiate in between a "base" set of
bindings which always applies regardless of the "current real"
context, and a "default" context which is already used as
a specialized context when no other specialized context (currently
only "compose") is active, in order to avoid that bindings remain
in, e.g., "compose" which make no sense there, like ":kf1 !date"
or the like -- it should be ":kf1 ~!date" there!

While here, disallow reverse solidus as the last character of
expansions, in order to avoid that newline escaping somehow comes
into play.