MLE: tab expansion: fix poss off-by-1; improve user experience..
To be able to get "the real feeling" our shell parser would need to
be able to split the command line into something like a DOM structure,
so that we would now _exactly_ what "states" of quotation etc. exists
at the point the cursor relies.  That we do not have.
But nonetheless, if i type "environ set TMPDIR=/<TAB>" then i want this
to expand to /tmp, i do not want to go back, split the shell token via
a separator, go back, expand, go back, unite