(BWDIC!) Only rm sysboxes, unless *posix* (Walter A. Iglesias, trondd)..
I think it is pretty weird behaviour to implicitly remove empty
mailbox files, especially so since we don't remove empty Maildir
etc. boxes, but only MBOX files.

The old Heirloom code had an *emptybox* variable which restricted
the standard *keep* variable for secondary and MBOX files, making
that responsible solely for system boxes.  But that is not
standard conformant, so i obsoleted it in, i think, early 2015.

Coming back to this after seeing a thread of Walter Alejandro and
trondd on misc@openbsd.org, i think the only sane behaviour is not
to remove any file by default.  However, we need to for system
boxes, so let's just keep the *keep*<->sysbox relationship, but
never remove any other MBOX mailbox unless *posix* is set, in
which case *keep* will also cover those.

In v15 we should have "VFS" and should be able to cover also other
mailbox types.  Maybe at that time we will add a *keep-other* or
whatever to reintroduce explicit behaviour, but the other way
around, as in *rm-secondary*, would be better.  Or such.