Final step to support Base64 with multibyte charsets..
Until now send.c:sendpart(): enabled iconv(3) for FROMB64_T, in
order to hook in the intermediate buffer that provides, so in
order to avoid that multibyte characters that cross Base64 buffer
boundaries get garbled by the makeprint() and delctrl() etc. we
apply to the data before showing it on the users' screen.

In v15 we will have a filter chain that uses buffers as
appropriate, until then hack the support in and keep lingering the
buffer around unless we see a NL (or end of message) before
performing further modifications and dumping out the data,
therefore keeping multibyte sequences (and even charset state
machines) intact.