Don't use replacement character for Base64 decoding..
ahem, just like we do for QP, because an iconv(3) may of course be
applied to the decoded data.
Just like stated in (More or less completely rewrite mime_enc.c..,
2016-11-04) i really dislike not giving users any feedback, but
what else can be done about this today.

While here, rename {b64,qp}_decode_text() to _part().
We are not yet there.