Add *r-option-implicit* (Felipe Gasper, Martin Neitzel)..
Felipe Gasper (felipe at felipegasper dot com) asked on
nail-devel@

  mailx sends mail using the “from” address I specify in
  ~/.mailrc, but when the message fails it won't send the
  rejection to that address.

  It *does* send the rejection as expected when I give an address
  with -r.

  Is this by design? It seems like a bug; the docs imply that
  setting “from” in ~/.mailrc should be a substitute for the -r
  flag.

Whereas the S-nail documentation is a bit give some notes
regarding this, and add a new option to perform the automatic
*from* -> -F/-f deduction without requiring the user to specify -r
explicitly (as the empty string or so) each and every time.

Then Martin stepped and that resulted in a huge improvement
regarding the manual, we now refer to RFC 5531's reverse-path etc.