Integrate S/MIME passwords into usual credential lookups (a bit)..
Until now we handle this very badly, and it can hardly be assumed
that anyone uses encrypted private keys in conjunction with us.

To top the horror, the password prompt -- which until now was the
sole thing to enter PEM passwords -- might appear while the $PAGER
is opened in a pipe, and we don't get this right, taking care for
the signals involved etc., that is to say.

To overcome this a bit, integrate PEM password lookup for S/MIME
signing and decryption into our usual (v15-style) credential
lookup by creating an URL with the given address and a pseudo
protocol, named ccred://, a.k.a. CPROTO_CCRED.
That is, we take email address and append pseudo-hosts,
.smime-cert-key, or .smime-cert-cert, or .smime-include-certs, and
perform a credential lookup for, e.g., bob@exam.ple.smime-cert-key.

So now all the possibilities for automatization of credential
lookup become possible for S/MIME PEM, too, mostly (encrypted)
.netrc lookup or (encrypted) resource file storage!