Extend range of per-message, a.k.a., compose-mode `localopts'..
the new hooks make only real sense if things like *customhdr*, and
all settings which can make up an account in general, message
signing etc., can be adjusted from within a hook and actually
affect the message.

This currently was not the case because we have unrolled before
leaving the collect() function -- still late enough to adjust
addressees, set signatures etc., but much to early for the
mentioned adjustments.

The reason was of course signals and our jumping around.  Since we
keep that barrel burst that the failed Sigman attempt is, let's
just use it for one more occasion, and in mail1(), in order to
delay calling temporary_unroll_compose_mode() to gain the desired
effects.