When displaying, avoid ugly header line linebreaks..
Because of our still present input==output line relationship we
are doomed to display Content-Transfer-Encoding caused line
breaks, which then sometimes looks so odd that a normal user
surely doesn't understand what is going on!
Since we cannot do anything about that for real, at least join it
and let it be folded by the user's display.