Security fix: never fall back to getenv(3)..
when looking up variables by string, unless explicitly allowed by
caller and only if the variable is not one that has an entry in
enum okeys.  This rather affects the shexp shell evaluation code
and the if/elif command only, since only those called getenv(3) on
their own if the normal variable lookup failed, but this is of
course a security flaw if the variable is a fully managed one,
because it introduces a bypass that is definitely unwanted.