The lunatic is on Solaris..
When i have implemented *on-compose-done-shell* i found the
necessity to redirect STDIN/STDOUT of the main process for the
duration of the hook.  Yes, that was because of our restricted
input handling, which should instead be event-based and emit
OnReadLineCompletedEvent so that consumers can attach.
And, of course, because there is no fdreopen() in STDIO.

But for a moment i also thought i should already (..because v15
will ship with its own, non-STD, I/O layer) provide n_{stdin,
stdout,stderr}, also rename the I/O functions already, and solely
use that.  I refrained from doing so after doing a grep on the
source.

Today, however, i first checked on Solaris since about October
2016, and our current approach is simply a no-go there, because
std{in,out,(err)} or not pointer based, but are macros which
expand to array indexes, i.e., instances.  Autsch.

So we do have to do the rename session of at least the variables
today, get rid of putchar() and plain printf(3).  Sigh.