Improve *expandaddr*=failinvaddr even interactive, e.g., with ~e..
HACK!  In pre-v15 we have a control flow problem (it is a general
design problem): if collect() calls makeheader(), e.g., for -t or
because of ~e diting, then that will checkaddr() and that will
remove invalid headers.  However, this code path does not know
about keeping track of senderrors unless a pointer has been passed,
but which it doesn't for ~e, and shall not, too.  Thus, invalid
addresses may be automatically removed, silently, and noone will
ever know, in particular not regarding "failinvaddr".

The hacky solution is the new bit EAF_MAYKEEP -- which can ONLY be
used for fields which will be subject to namelist_vaporise_head()
later on!! --, if it is set (by extract_header()) then checkaddr()
will NOT strip invalid headers off IF it deals with a NULL
senderror pointer