Drop requirement of BYTE_ORDER knowledge ++ for OPT_CROSS_BUILD..
I hate to say it but this is a program and when we want to support
cross building for a program that noone really cares about that
much (as of today), then we must simplify our environment.

(This changeset first implemented a new yrun() mk-conf.sh function
which parsed a new VAL_CROSS_BUILD make.rc variable which could be
used to specify via +FEATURE / -FEATURE (or left off) whether
something was true or not, but after thinking about it the
+little_endian that we have needed appeared to be overly silly.)

What i really hate is that now our bit fields are no longer layed
out the way i want to, but i guess it is not really that much of
a problem for our use cases.  And big endian machines are not so
common anyway in the end.  In the world i live in, that is to say.