Oh no, nonono: *0* -> *!*, better that is..
Using $0 for the (soft) return status seemed to be a good idea
because the meaning it has in the shell will never be used for us.

However, now that `vexpr' exists and supports regular expression
search+replace we actually have a real use case for it, and this
one is known by the people for many, many years!

So instead of inventing more and more unusual things, use *!* as
the variable backing the (soft) return status, the meaning it has
in the shell will never be an issue for us either, and then it
seems to be "nearer" to *?* than *0* ever was.

And, while here, support *0* for the overall regular expression
match (of `vexpr'), what it is ment to be used for.