nail.1: clarify/FIX doc. of \cX (Noel Chiappa, Random832)..
Whereas our code is correct, it is absolutely true that the
documentation was overly simple.  In fact it was a false lie.
To detect the ASCII character that represents a control code, you
do not add 64, but you xor it.

Many thanks to Noel Chiappa (jnc at mercury dot lcs dot mit dot
edu) and Random832 (random832 at fastmail dot com).