Extend shell argument parsing and quoting compatibility..
We yet did neither perform proper quoting at all, and we also did
not support meta control characters.
For now implement ;, but be aware of & and |, too.  We will never
support parsing of (, ), < and >, but we will quote them, for the
sh(1)ell.

So now this is getting as cool as normally expected:

  localopts yes; wysh set verbose; ignerr eval "${@}"; return ${?}

Works.  _If_ `set' is lifted towards future via `wysh', so that
new-style parsing rules are used for the remainder of the line.