(BWDIC!) Redefine all CODECS, add `shcodec', change `addrcodec'..
The thing about codecs is that a user shall be enabled to feed in
data and get back the encoded and/or decoded variant of that data
according to whatever the codec is about.

Whereas we now have shell-quoting and could tell the user that the
input needs to be a valid shell-input string, that mechanism (1)
requires the knowledge that the `shcodec' is about to provide,
i.e., the user has to know about it from the very start, and (2)
i really think that "codec"s are at times a very difficult thing,
and it is much better if the user simply types in exactly what is
desired, and the codecs simply makes the best of it!

So redefine codecs, and use TYPE_RAWDAT input for all of them.