Add n_lex_input_clearerr(); limit *ignoreeof* count to 4..
strace(1) on Linux mistreats our FDs on interrupt (or maybe it is
doing just as weird on signals as we do), and we could be forced
to an endless *ignoreeof* loop.  Not in collect(), there it is
already limited, however.  But in general it makes no sense to
offer *ignoreeof* if the EOF and error states are not cleared
first, so add that function.