(BWDIC!) FIX false decision: for `if' etc., add -le,-gt.. operators..
It was a dumb and false decision to perform automatic integer
conversion for ==, != etc. comparison operators, so that 0==0x0 is
true, but which you don't want if both shall be strings.

Therefore go the shell test(1) way and support the -le etc.
operators, use them instead for integers, giving error for
non-integer arguments, and let the == etc. series be exclusive for
strings.

While here, some style updates.