No underscores in files names (_only_ hyphen-minus)..
Yes, until now we had a mixed usage.  This was because
C names cannot have hyphen-minus in them, and we need the n_FILE
macro for compile-time-assertion expansions etc., e.g.,

  typedef int n_CONCAT(avoid_empty_file__, n_FILE);

So this would have been a no-go.  However, let's ignore the fact
that in such messages the filename will no longer match the real
file name, but have instead ls(1) give a unique view.