(BWDIC!) No!  Let $! be errno, and $? the sole exit status..
Making a difference in a "soft" and a "hard" exit status because
of state machine restrictions is a screaming mess.  No.

Instead we have introduced our own errno(X) wrapper machinery in
(Our own layer onto errno, yet too high), and now turn to use it
and report "the errno" via $! (and as well in $^ERR, $^ERRNAME,
$^ERRDOC).

Adjust the state machinery to allow more exit statuses than (-1),
0 and 1, and let $? thus be _the_ exit status.
Introduce the new *errexit* variable which acts the same as what
"set -e" does in a POSIX sh(1)ell, obsolete *batch-exit-on-error*,
which just maps to *errexit* until it is dropped.
Add a new $POSIXLY_CORRECT/*posix* entry and exit the program if
an error occurs during loading of resource files during startup.