Use autorec memory for the ARGV of commands..
and reduce the size of maximum ARGV so that it really fits and
doesn't go over (yet) system allocator.
The problem was that in code

  define work {
    eval echo $1: \$^ERR-$1: ERR-\$^ERRNAME-$1: \$^ERRDOC-$1
    vput vexpr i + $1 1
    if [ $i -lt 1000 ]
      call work $i
    end
  }
  call work 0

we crashed because of stack exhaustion (over 4000 frames of stack
deep), and this is one heavy consumer we can get rid of easily.

TODO Of course this is all intermediate in that the ARGV must be
TODO part of a CmdCtx structure, and of course this can be
TODO dynamically created to just fit for the actual needs!
TODO But not yet, unfortunately.
TODO Due to lack of loops i am thinking about an `exec' or `xcall'
TODO feature which simply replaces the current macro context with
TODO that of the called one, which could possibly become
TODO implemented somewhat easily, and then a user could loop
TODO endlessly as above, and not cause a SIGSEGV after fewest
TODO thousands iterations!