C****Y BSD signal handling!  Try improve ^C stability..
It is all my fault.  I should have sat back and think about it,
then deal with the BSD SA_RESTART signal handling that is used
by blocking all signals and search those places in the code where
we have to deal with them, and use as-local-as-possible condomes.
If i would have done this back in 2013, we would look better today

So let's reorder things a little bit, and extend the
hold_all_sigs() ranges around lot of our management code, which
significantly reduces crashes while running an endless loop via
`xcall'.  Of course it doesn't help at all, since we potentially
leave external libraries in undefined state, and... sigh... let's
just do this now.  It improves the things a little bit.

While here, move data to a new n_go_data context carrier, which we
manage as part of our evaluation context handler, and which will
grow over time.  Like this, memory.c and colour.c do not longer
need to special treat internal data, but simply use that of the
current context.  Especially colour is still bad since it should
be object based, not forming a real stack, but it is a step.