Popen(): ignore SIGINT if n_CHILD_FD_PASS..
Ooh, we yet did only do so for n_child_run(), but not for
normal Popen()..Pclose().  Fix a possible leak while here.

P.S.:
this is all terrible -_-
We need to track once the first command is started with takes over
the terminal, and restore once the last such is finalized.
Process groups etc., like in TODO (or my brain) since years, and
YEARS to come