Improve shell parser sequencing; changes \c@..
The skip mechanism was (also) not adjusted after input sequencing
via metacharacters has been implemented, and continued to skip over
anything.

This also changes the meaning of \c@: whereas it yet stopped
argument processing entirely, it now only stops recognition of the
current token, or, dependent on the context, of all the remaining
arguments of the command.  That is still different to
\0/\x0/\u0/\U0, which extend to the end of the current quote only.
Before

  ?0!0[#/var/spool/mail/steffen]? echo $'hey\c@\07777888a'du;echo hu
  hey
  ?0!0[#/var/spool/mail/steffen]? echo $'hey\c@\07777888a'du bla;echo hu
  hey

After

  ?0!0[#/var/spool/mail/steffen]? echo $'hey\c@\07777888a'du;echo hu
  hey
  hu
  ?0!0[#/var/spool/mail/steffen]? echo $'hey\c@\07777888a'du bla;echo hu
  hey
  hu