Extend $SOURCE_DATE_EPOCH coverage to randoms+..
A problem of reproducible-builds.org is that the only indication
that a reproducible environment is desired is $SOURCE_DATE_EPOCH.
Anyhow, that is how it is.

Extend reproducibility to complete coverage (?) by using
deterministic counters instead of randoms for Message-ID:s etc.,
a (hopefully non-existing) $LOGNAME, a unique version number
(except for -v and `version') etc. when we detect upon startup
that $SOURCE_DATE_EPOCH is set.  This setting in non-reversible.