Readd removed fflush()/fseek() in between read and write..
The behave:record_a_resend-2 test failed on Solaris 5.10 and 5.11,
messages in *record* where separated by four \n / U+000A instead
of two.  It turns out that the effective sequence

   if((fo = Zopen(name, "a+")) == NULL){
      if((fo = Zopen(name, "wx")) == NULL){

    ...

      if(fseek(fo, -2L, SEEK_END) == 0){
         switch(fread(buf, sizeof *buf, 2, fo)){

    ...
         if(emptyline){
            putc('\n', fo);
            fflush(fo);
         }

on a O_APPEND|O_RDWR file requires an intermediate

  fseek(fout, 0, SEEK_END))
or
  fflush()

which i consider a bug in Solaris stdio.
Jörg Schilling however convinced me that the standard requires
this behaviour, and it is also in the C FAQ.  Nonetheless...

While here, unite the two existing use cases of such MBOX append
preparation in a new function n_folder_mbox_prepare_append() as
long as all this is not as "simple" as

  URL *urlp = URL::from_string(DATA);
  Mailbox *mbp = Mailbox::create_from_url(urlp);
  mbp->append()

and add a new use case for file addressees, as that yet simply
appended to yet existing files instead of checking whether a valid
MBOX database is present or true separation is needed.