Add OPT_ASAN_{ADDRESS,MEMORY} options..
Hm, the AddressSanitizer helped to find leaks in xssl.c (rebased
back as (cc-test.sh: one more checksum for S/MIME test)), but the
MemorySanitizer of clang 4.0.0 (ArchLinux 4.0.0-2) seems to be
faulty.  Also it just can't seem to make interception work out,
but anyway, the idea is very smart, and those leaks in xssl.c (of
which i know, there is (still) a TODO note on leaks) were not
reported by valgrind.