Add enum n_fopen_state, rename Zopen()->n_fopen_any(), etc..
Rather an add-on to (Support protocol:// prefix to fixate++
`file'++ type).  We have this utterly ugly code all around there
which does Zopen("a+") followed by Zopen("wx") just to see whether
a file did yet exist, and of course _that_ "a+" is not standard
conform at all, because there "a+" includes O_CREAT.

So pass some status back on request; in fact we should return
the state and store a FILE* result in a FILE** arg, but in the
future this will be all dropped and replaced by a factory() in
a Mailbox multiplexer.

More cleanup to come.