Adjust inject/history handling for new, *ifs* trimmed input..
Introduction of *ifs* blew our simple-minded history
handling, because we yet simply placed a space at the front to
avoid this, also for (some) key-`bind'ings.
This changeset improves this situation quite a bit.
It is far from being ok, which would be passing a context carrier
all along the code path, with whoever thinks he nows better
setting CTX->history_add to FAL0.
But anyway, it is a bit better now.