Various fixes to actually integrate IFS for at least some for real..
Using the new (still incomplete and only supporting flat table)
option parser to be able to truly support IFS has had the bad
side effect of requiring that we use RAWDAT argument type in the
normal command table, and that made actually the `xcall' test
fail: of course, because the macro data environment which holds
the local arguments which we want to expand when we say

  xcall BLA "$@"

are already gone!  That is the essence of the `xcall'
optimization, and did not hurt with WYSH parsing because there the
arguments became expanded before c_xcall() has been invoked.

To overcome this, we have to incorporate the new option parser
into the normal command table and add a new command argument type
for it, so that the evaluation takes place at the same time as it
does for WYSH.  Sigh, i explicitly avoided doing this work once
i have implemented `bind' and `unbind', the first users of the new
parser, but now it happened.

It made thus sense to reorganize the code a little bit more,
outsource the command table into its own file (again), and merge
cmd-arg.c with it.  At the cost of needing to expose some formerly
static functions, of course.  I thought that was worth it.

TODO We are far from being there, the command table should be
implemented as a multi-dimensional array where the first level is
indexed by first letter of command MODULO size, and all that
should be sorted alphabetically excluding POSIX standardized
abbreviations ordering first.  All that should be extracted
automatically from especially crafted comments in the sources, in
or at the function definitions.

Also, unrolling `commandalias' code in go.c was fine but besides
these are normal key/value tuples and thus belong into nam-a-grp.c
where all the other hashtable stuff is located.