`~^ header': add `remove-at', test in t_behave_compose_hooks ++..
and let the `insert' subcommand return the index of the newly
inserted header...

Split the huge a_collect_plumbing(), finally these really need to
be table-based.

Change elide(), namelist_dup() and usermap(): it always hurt me
that we change the order of the user-specified headers, because we
perform a list sort to strip off duplicates.  Instead, create
a temporary array, qsort() that, and the walk the list and check
whether a node is (still) contained in the array.  Improvements
are possible.