n_shexp_parse_token(): fix PARSE_META_SEMICOLON trail WS strip..
I hate to say it.
We are not the shell, we are not a language, and therefore the
general *ifs-ws* and normal whitespace trimming that input lines
undergo (in a_go_evaluate()) has already happened before we enter
the shell token parser, and so the result will be used *as is*,
and therefore it may occasionally be necessary to be aware of and
remove trailing unquoted WS that would otherwise remain, after we
have seen a semicolon sequencer.