Change *on-compose-{enter,leave}* environment + fixes..
Those simple macro-based hooks will henceforth (and likely
temporarily only) get access via *mailx-{subject,to,from...}*
instead of *compose-{..}*.
This is because we will get more hooks and then those may not
really run on compose mode.

Extend the set of headers which are injected: we now keep track of
the very original headers, before any alias expansion, name
munging, whatever, as Mailx-Orig-{To,Cc,Bcc}, and these are also
made available to mentioned hooks.
They can be accessed (readonly) via ~^header, and while doing so
a bug was discovered that was not covered by test output.  (Yes,
tests do not make sense if a blind worm verifies their output.)

Our codebase is not yet able to truly fill in all those headers
all the time, because we sometimes not even parse the message and
thus don't know them.  This is not for performance though, since
we may parse them later nonetheless, but without knowing.  In v15
we will slurp it all anywhere just to create a DOM, because we
really require knowledge about the headers.  Then -t will also
work, no matter what there is.  But all this one more step.