Better hook support: *mailx-{raw,orig}-* a.k.a. Mailx-{Raw,Orig}-From:..
The hooks need more context to be useful, so -- even though
incredibly expensive as of today, hopefully _much_ less so in v15
-- regulary provide a set of temporary, virtual headers which
transport header values of the messages being
replied-to/forwarded/resent via Mailx-Orig-*: a.k.a
*mailx-orig-**, and of the current message before any name manging
etc. (except for alias processing) via Mailx-Raw-* a.k.a.
*mailx-raw-**.

While here, obsolete *fwdheading* in favour of
*forward-inject-head*, obsolete `Fwd' and `fwd', they are aliases
of `Forward' and `forward', respectively, and have no equivalent
anymore in, e.g., `headerpick', etc.

Support memory relaxation for `resend' series, too, just in case.

This fixes quite some test errors.  Errors in the test that is. ;)