FIX [62358d22] (More `alternates' processing: a 2nd "pure"..run)..
Bummer!  On a fuzzy afternoon [62358d22] (More `alternates'
processing: a 2nd "pure", non-extended run) added more stripping
but for one counts non-`alternate's as if they were `alternate's,
and second does not terminate the tail of a newly relinked list!
What a terrible mess!

The approach as such was brain damaged.  Instead simply mark all
messages which refer to an `alternate' in order to finally (!)
remove all such marked messages (instead of possibly one).
That shall be it.